Fix non-functional S3 wakeup with forced empty HID report
This bug was introduced with the de-duplication of USB-HID reports in
8b865a9d6445e3bce6c991224f373d71c67e87e5 by yours truly. The problem was
that the wake-up report (an empty report most of the time) wasn't send
if it matched the previous report before suspension.
Interestingly this bug only surfaced on keyboards without any additional
endpoints, e.g. console output, that would send a report once the main
loop resumed.

The bug is fixed by forcefully sending an empty report after wakeup to
trigger a resume of the host.